install: rename prepublish.js to prepare.js
PR-URL: https://github.com/npm/npm/pull/14290
Credit: @othiym23
