+++ lib/install.js
@@ -549,7 +549,7 @@ Installer.prototype.runPostinstallTopLevelLifecycles = function (cb) {
     [doOneAction, 'postinstall', this.idealTree.path, this.idealTree, trackLifecycle.newGroup('postinstall:.')])
   if (this.dev) {
     steps.push(
-      [doOneAction, 'prepublish', this.idealTree.path, this.idealTree, trackLifecycle.newGroup('prepublish')])
+      [doOneAction, 'prepare', this.idealTree.path, this.idealTree, trackLifecycle.newGroup('prepare')])
   }
   chain(steps, cb)
 }
