+++ lib/install/action/prepare.js
